requirements.txt
a2wsgi==1.10.8
ansio==0.0.1
beautifulsoup4==4.12.3
browser-use==0.1.40
# docker==7.1.0  # Optional - only needed for isolated code execution
duckduckgo-search==6.1.12
faiss-cpu==1.8.0.post1
fastmcp==2.3.4
flask[async]==3.0.3
flask-basicauth==0.2.0
flaredantic==0.1.4
GitPython==3.1.43
inputimeout==1.0.4
langchain-anthropic==0.3.3
langchain-community==0.3.19
langchain-google-genai==2.0.8
langchain-groq==0.2.2
langchain-huggingface==0.1.2
langchain-mistralai==0.2.4
langchain-ollama==0.2.2
langchain-openai==0.3.1
# openai-whisper==20240930  # Optional - only needed for speech-to-text

# MongoDB Atlas dependencies
pymongo==4.6.1
motor==3.3.2
langchain-mongodb
lxml_html_clean==0.3.1
markdown==3.7
mcp==1.9.0
newspaper3k==0.2.8
paramiko==3.5.0
playwright==1.52.0
pypdf==4.3.1
python-dotenv==1.1.0
pytz==2024.2
# sentence-transformers==3.0.1  # Optional - only needed for local embeddings
tiktoken==0.8.0
unstructured==0.15.13
unstructured-client==0.25.9
webcolors==24.6.0
nest-asyncio==1.6.0
crontab==1.0.1
